move packet module to valence_network
